requirements.txt
arcade>=3.0
Pillow>=10.0
